apps/model-deployment/requirements.txt
gradio
huggingface-hub
llama-cpp-python
